pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lltune"
version = "0.1.0"
description = "Low-Latency System Tuner CLI and library"
requires-python = ">=3.9"
authors = [{name = "Muhammad Nawaz", email = "m.nawaz2003@gmail.com"}]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
keywords = [
  "low-latency",
  "latency",
  "HFT",
  "high-frequency-trading",
  "trading",
  "linux",
  "performance",
  "tuning",
  "optimization",
  "kernel",
  "network",
  "cpu-isolation",
  "irq-affinity",
  "hugepages",
  "numa",
  "real-time",
  "sysctl",
  "server",
  "fintech",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: Financial and Insurance Industry",
  "Intended Audience :: System Administrators",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: System :: Systems Administration",
  "Topic :: System :: Networking",
  "Topic :: System :: Operating System Kernels :: Linux",
  "Typing :: Typed",
]
dependencies = [
  "pyyaml>=6.0",
  "ruamel.yaml>=0.17",
]

[project.optional-dependencies]
dev = [
  "pytest>=7",
  "flake8>=7",
  "build>=1",
  "twine>=5",
]

[project.scripts]
lltune = "lltune.cli:main"

[project.urls]
Homepage = "https://github.com/nawaz1991/lltune"
Repository = "https://github.com/nawaz1991/lltune"
Documentation = "https://github.com/nawaz1991/lltune#readme"

[tool.setuptools.packages.find]
where = ["src"]
